fix: the same bit_width narrowing in the pair recertifier test
Second and last site of the width-type difference between libstdc++
versions. Same reasoning: the value is at most 64, so the cast is exact.

+++ morsehgp3d/tests/unit/test_anchored_pair_candidate_recertifier.cpp
@@ -237,7 +237,7 @@ constexpr std::uint64_t kFnvPrime = UINT64_C(1099511628211);
     const std::uint64_t difference =
         leaves[begin].morton_code ^ leaves[end - 1U].morton_code;
     if (difference != 0U) {
-      const int width = std::bit_width(difference);
+      const int width = static_cast<int>(std::bit_width(difference));
       const unsigned int highest_bit = static_cast<unsigned int>(
           width - 1);
       const std::uint64_t mask = UINT64_C(1) << highest_bit;
